# audit: make audit_revision_and_version_scheme faster.