@MikeMcQuaid 2020-06-05
@Bo98 Have added a `--git` (PR to add it to test-bot incoming). Let me know what you think. I'm aware some tests are failing.